LPS-91267 Move BytecodeJavaFileObject out of JspCompiler
